PR리뷰 반영 - lottos로 winningLotto를 보내서 결과물을 만들어 내도록 구조 변경 https://github.com/next-step/kotlin-lotto/pull/799#discussion_r1257314502
