+++ src/main/kotlin/lotto/domain/Lottos.kt
@@ -3,4 +3,11 @@ package lotto.domain
 class Lottos(val lottos: List<Lotto>) {
     val size: Int
         get() = lottos.size
+
+    fun match(winningLotto: WinningLotto): WinningStatistics {
+        return lottos.map { lotto -> winningLotto.rank(lotto) }
+            .groupingBy { it }
+            .eachCount()
+            .let { WinningStatistics(it) }
+    }
 }
